refactor: address ulong-migration review feedback
Cleanups across the long->ulong migration:
- Simplify saturating subtractions via SaturatingSub (DbBlocksLoader,
  BlockAccessListManager validation) and replace float-ratio memory math
  with integer division in MemoryHintMan; divide by 1UL.MB instead of /1000/1000.
- Remove redundant/awkward casts: BLS MSM/pairing gas (uint->ulong),
  Eip2537 discount table to ulong, IntrinsicGasCalculator tests,
  OptimismTransactionProcessor ValueRef, BlockchainBridge nonce (int overload),
  EvmPooledMemory MaxCodeSize, MemoryHintMan/MemoryHintManTests.
- Era readers: read block count/start number as UInt64 (matching original),
  EraWriter writes UInt64; EraE _blockCount/BlockCount to ulong.
- LogIndexStorage _maxReorgDepth kept int (block numbers are int by design there).
- HeaderValidator: rename ValidateAbsoluteGasLimit->ValidateFieldLimit, run first.
- Add Random.NextUInt64 (with bound overloads) and a shared TestRange.ULongRange
  helper; use across tests instead of (ulong)random.Next()/Enumerable.Range.
- BlockhashCache.Stats: int ctor to avoid production casts.
- BatchV1: restore one-line tuple deconstruction in the three tx decoders.
- Eth70: name the no-block-seen sentinel; DebugRpcModule one-liner;
  Eip1108/Eip152 block-number offset via shared helper.
- Drop noise comments (TransactionBuilder, SimulateBridgeHelper, HistoryPruner
  banner, E2StoreReader, ReceiptMigration); HistoryPruner uses Rlp.Encode(ulong).

Also fixes pre-existing Xdc.Test compile breakage from the ulong sign-window
change (MergeSignRange/MinePeriod/GetSnapshotByBlockNumber).

+++ src/Nethermind/Nethermind.Db/LogIndex/LogIndexStorage.cs
@@ -184,7 +184,7 @@ namespace Nethermind.Db.LogIndex
 
         private readonly ILogger _logger;
 
-        private readonly ulong _maxReorgDepth;
+        private readonly int _maxReorgDepth;
 
         private readonly AllMergeOperators _mergeOperators;
         private readonly ICompressor _compressor;
@@ -228,7 +228,7 @@ namespace Nethermind.Db.LogIndex
             {
                 Enabled = config.Enabled;
 
-                _maxReorgDepth = config.MaxReorgDepth!.Value;
+                _maxReorgDepth = (int)config.MaxReorgDepth!.Value;
 
                 _logger = logManager.GetClassLogger<LogIndexStorage>();
 
@@ -512,7 +512,7 @@ namespace Nethermind.Db.LogIndex
             return (min, max);
         }
 
-        private int? GetLastReorgableBlockNumber() => _maxBlock - (int)_maxReorgDepth;
+        private int? GetLastReorgableBlockNumber() => _maxBlock - _maxReorgDepth;
 
         private static bool IsBlockNewer(int next, int? lastMin, int? lastMax, bool isBackwardSync) => isBackwardSync
             ? lastMin is null || next < lastMin
